\name{GunmaPop2005}
\docType{data}
\alias{GunmaPop2005}
\title{Census data of Gunma Prefecture, Japan in 2005}
\description{
 The data gives the populations of Gunma Prefecture, Japan
 for each age of 0 to 108, males and females separately
 according to the population census 2005.
}
\usage{ GunmaPop2005 }
\format{ A data frame with 109 observations and 3 variables. 

 \tabular{rlll}{
  \code{[, 1]} \tab \code{Males} \tab numeric \tab Age specific population of males in 2005 Gunma Pref.\cr
  \code{[, 2]} \tab \code{Females} \tab numeric \tab Age specific population of females in 2005 Gunma Pref.\cr
  \code{[, 3]} \tab \code{Ages} \tab numeric \tab Ages from 0 to 108.\cr
 }

}
\source{ \url{http://www.e-stat.go.jp/SG1/estat/List.do?bid=000001005042&amp;cycode=0}. }
\keyword{datasets}
